refactor: drawer window.event 전역 사용
