+++ src/components/share/ShareDrawer.tsx
@@ -53,17 +53,6 @@ const ShareDrawer: React.FC = () => {
     });
   }, [api]);
 
-  useEffect(() => {
-    const handlePopState = () => {
-      setIsOpenShareDrawer(false);
-    };
-
-    window.addEventListener("popstate", handlePopState);
-    return () => {
-      window.removeEventListener("popstate", handlePopState);
-    };
-  }, [isOpenShareDrawer, setIsOpenShareDrawer]);
-
   const handleDotsClick = (index: number) => {
     if (!api) return;
     setCurrentIndex(index);
